Move test constructs out of src file
